pyproject.toml
[tool.poetry]
name = "tosslate_flask"
version = "1.0.1"
license = "MIT"
authors = ["Your Name <you@example.com>"]
description = ""

[tool.poetry.dependencies]
python = "^3.8"
arrow = "^1.2.2"
Authlib = "^1.0.1"
Cerberus = "^1.3.4"
click = "^8.1.3"
Flask = "^2.1.2"
Flask-Bcrypt = "^1.0.1"
Flask-Limiter = "^2.4.5"
flask-marshmallow = "^0.14.0"
graphene = "^3.0"
gunicorn = "^20.1.0"
peewee = "^3.14.10"
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]
bpython = "^0.22.1"
rich = "^12.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
